Add files via upload
This notebook is part of Tokyo Data Science summer 2020 by Professor Michal Fabinger. 

Task 1
The Numpy code in the notebook assumes that the Huber loss threshold value is 1. Modify the code to make it work with a general Huber loss threshold. Name that variable huber loss threshold for clarity. Then perform the calculation in your notebook setting huber loss threshold = 0.6.

Task 2

The PyTorch code in the notebook also assumes that the Huber loss threshold value is 1. Modify the code to make it work with a general Huber loss threshold. Keep the same name huber loss threshold for clarity. Then perform the calculation in your notebook setting huber loss threshold = 0.6. Confirm that you are getting the same results as in Task 1 if you run sufficiently many gradient descent steps with a small learning rate. If there is a significant discrepancy, you should
search for mistakes in your code.